Moved notes page components to new files. Reorganized file structure
